Add option to keep files in S3 when publishing
Add a '--keep-files' option to the publish management command
that causes the command to skip the deletion of files that
have keys in the S3 bucket, but don't exist in the build directory.

This is a simple way to support safer publishing in cases where
items may be written to the S3 bucket outside of the Django
Bakery build/publish workflow.  With this option, users can
publish their Django project with Django Bakery and have the
flexibility to manage other components that publish to the
same S3 bucket outside of Django.

To test this feature, use [moto](https://github.com/spulec/moto)
to mock the boto library, used to interact with S3.  The mock
package is also needed to work around some weirdness with the
S3 mocking that isn't crucial to the feature under test.

Update setup.py to include the testing dependencies.